Clean up use of global state/bindings
This patch takes advantage of the novel concept of wrapping the routes
in a fn which serves to close over the app config as an immutable
value. This allows the routes to directly reference destructured
configuration values and for me to do away entirely with the previous
structure of dynamic variables.